requirements.txt
journal_api~=1.3.8
requests~=2.32.3
aiohttp~=3.10.5
beautifulsoup4~=4.12.3
aiogram~=3.12.0
python-dotenv~=1.0.1